requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.1
colorama==0.4.6
fastapi==0.135.3
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
SQLAlchemy==2.0.48
starlette==1.0.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.42.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0requests==2.32.3